pyproject.toml
[project]
name = "tds-p2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "beautifulsoup4==4.12.2",
    "camelot-py[cv]==0.11.0",
    "chardet==5.2.0",
    "chromium>=0.0.0",
    "duckdb>=1.4.2",
    "fastapi==0.104.1",
    "folium==0.15.1",
    "fuzzywuzzy==0.18.0",
    "geopandas==0.14.1",
    "geopy==2.4.1",
    "html5lib==1.1",
    "httpx==0.25.2",
    "jsonschema==4.20.0",
    "kaleido==0.2.1",
    "loguru==0.7.2",
    "lxml==4.9.3",
    "matplotlib==3.8.2",
    "networkx==3.2.1",
    "numpy==1.26.2",
    "openai>=1.6.1,<2.0.0",
    "opencv-python==4.8.1.78",
    "openpyxl==3.1.2",
    "pandas==2.1.4",
    "pdfplumber==0.10.3",
    "pillow==10.1.0",
    "playwright==1.40.0",
    "plotly==5.18.0",
    "polars==0.19.19",
    "pyarrow==14.0.1",
    "pydantic-settings==2.1.0",
    "pydantic[email]==2.5.0",
    "pypdf==3.17.0",
    "pypdf2==3.0.1",
    "pytesseract==0.3.10",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "python-dateutil==2.8.2",
    "python-docx==1.1.0",
    "python-dotenv==1.0.0",
    "python-levenshtein==0.23.0",
    "python-multipart==0.0.6",
    "pytz==2023.3",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "scikit-learn==1.3.2",
    "scipy==1.11.4",
    "seaborn==0.13.0",
    "shapely==2.0.2",
    "sqlalchemy==2.0.23",
    "tabula-py==2.9.0",
    "tenacity==8.2.3",
    "toml==0.10.2",
    "tqdm==4.66.1",
    "uvicorn[standard]==0.24.0",
    "xlrd==2.0.1",
]
